Extract statments module
